% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runPipeline.R
\name{read}
\alias{read}
\title{Read User Input from Console}
\usage{
read(prompt = "")
}
\arguments{
\item{prompt}{string to display as prompt for user input}
}
\value{
Returns the user input as a string.
}
\description{
This function reads user input from the console with a given prompt. It is a
wrapper for the base R `readline` function. This function was created only
for ease of testing and mocking during unit tests.
}
\examples{
# No examples: this function requires user input

}
