% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runDenoising.R
\name{runDenoising}
\alias{runDenoising}
\title{Apply Denoising Algorithms To Tomograms}
\usage{
runDenoising(image, output_dir = NULL)
}
\arguments{
\item{image}{tomogram file to be denoised}

\item{output_dir}{optional directory to save the denoised images, otherwise
images will be saved in the data folder of the package}
}
\value{
Returns null, instead saves the denoised images.
}
\description{
This function performs image denoising on tomograms (in .tif format) using
several filtering algorithms implemented in the imager package. The supported
denoising algorithms include Gaussian blur, median filter, and anisotropic
diffusion filtering. The function reads and displays the input image, applies
the denoising algorithms, and visualises the results for comparison. This is
meant to assist the user in evaluating the effectiveness of different
denoising algorithms before passing them into segmentation.
}
\examples{
# Example 1:
# Running denoising on a sample tomogram

image <- TS_001.133
cryoCompare::runDenoising(image, tempdir())

}
\references{
https://cran.r-project.org/web/packages/ijtiff/index.html
https://cran.r-project.org/web/packages/imager/index.html
https://pubmed.ncbi.nlm.nih.gov/6102765/
https://www.sid.ir/paper/544308/en
https://www.mia.uni-saarland.de/weickert/Papers/book.pdf
}
