% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runPipeline.R
\name{runPipeline}
\alias{runPipeline}
\title{Run Complete Denoising and Segmentation Pipeline}
\usage{
runPipeline(image_path, methods, ground_truth_image_path = NULL)
}
\arguments{
\item{image_path}{path to the tomogram file in .tif format}

\item{methods}{vector of strings of segmentation methods to apply}

\item{ground_truth_image_path}{optional path to the ground truth tomogram
file in .tif format}
}
\value{
Returns null as this is the end of the pipeline
}
\description{
This function first applies several denoising filters to a tomogram in .tif
format, then prompts the user to select one of the denoised images (or the
original) for segmentation. It then runs the specified segmentation
algorithms on the chosen image. This function integrates the denoising and
segmentation steps into a single pipeline for ease of use.
}
