% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runSegmentation.R
\name{runSegmentation}
\alias{runSegmentation}
\title{Run Segmentation Pipeline On Tomograms}
\usage{
runSegmentation(image, methods, ground_truth = NULL)
}
\arguments{
\item{image}{tomogram file to be run through the pipeline}

\item{methods}{vector of strings of segmentation methods to apply; choose
from "Huang", "Mean", "Otsu", and "Triangle"}

\item{ground_truth}{optional ground truth mask of tomogram file}
}
\value{
Returns a data frame containing the DICE similarity coefficient for
each segmentation method compared to the ground truth, if provided. Otherwise,
returns NULL.
}
\description{
This function will run segmentation algorithms on the specified tomograms
(must be in .tif format). It is preferred that the tomograms are already
denoised, as segmentation algorithms typically perform better on denoised
data. The function supports multiple segmentation algorithms, including
various traditional image thresholding methods including Huang
(Huang & Wang, 1995), Mean (Glasbey, 1993), Otsu (1979), and Triangle
(Zack et al., 1977). This function will read the tiff tomogram file,
then send the results to other functions to apply the specified segmentation
algorithms, visualise the results, and compare them against ground truth
if provided.
}
\examples{
# Example 1:
# Segmentation of a tomogram with Otsu and Triangle algorithms without ground truth

data("TS_001.133", package = "cryoCompare")
image <- TS_001.133
cryoCompare::runSegmentation(image, methods = c("Otsu", "Triangle"))

# Example 2:
# Segmentation of a tomogram with all available methods with ground truth
data("TS_001.133", package = "cryoCompare")
data("TS_001.133_ground_truth", package = "cryoCompare")
image <- TS_001.133
ground_truth <- TS_001.133_ground_truth
cryoCompare::runSegmentation(image,
methods = c("Huang", "Mean", "Otsu", "Triangle"), ground_truth)

}
\references{
Glasbey, C. A. (1993). An analysis of histogram-based thresholding
algorithms. CVGIP: Graphical Models and Image Processing, 55(6), 532-537.
https://doi.org/10.1006/cgip.1993.1040
Huang, L., & Wang, M. J. (1995). Image thresholding by minimizing the
measures of fuzziness. Pattern Recognition, 28(1), 41-51.
https://doi.org/10.1016/0031-3203(94)E0043-K.
Nolan, R., & Padilla-Parra, S. (2018). ijtiff: An R package providing TIFF
I/O for ImageJ users. Journal of Open Source Software, 3(23), 633.
https://doi.org/10.21105/joss.00633
Otsu, N. (1979). A threshold selection method from gray-level histograms.
IEEE Transactions on Systems, Man, and Cybernetics, 9(1), 62-66.
https://doi.org/10.1109/TSMC.1979.4310076
Zack, G. W., Rogers, W. E., & Latt, S. A. (1977). Automatic measurement of
sister chromatid exchange frequency. Journal of Histochemistry &
Cytochemistry, 25(7), 741-753. https://doi.org/10.1177/25.7.70454
}
