% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runSegmentation.R
\name{thresholdSeg}
\alias{thresholdSeg}
\title{Apply Thresholding Segmentation Methods}
\usage{
thresholdSeg(image, methods)
}
\arguments{
\item{image}{tomogram image read into R}

\item{methods}{vector of strings of segmentation methods to apply}
}
\value{
Returns a list containing the mask and threshold values for each
method
}
\description{
This function applies specified traditional image thresholding algorithms to
a tomogram to generate binary masks for segmentation. The function takes
in an already read image, computes the individual masks for each method, and
returns them along with their respective threshold values for further
analysis. While this is a helper function for the main segmentation pipeline,
it can also be used independently for obtaining masks and thresholding values
for further analysis.
}
\examples{
# Example 1:
# Extracting the segmentation masks and thresholds using Otsu and Triangle methods
image <- TS_001.133
seg_result <- cryoCompare::thresholdSeg(image, methods = c("Otsu", "Triangle"))

}
\references{
https://cran.r-project.org/web/packages/ijtiff/index.html
https://cran.r-project.org/web/packages/autothresholdr/index.html
https://imagej.net/plugins/auto-threshold
}
