% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TS_001.133}
\alias{TS_001.133}
\title{Tomogram Of SARS-CoV-2 Spike Proteins On Intact Virions}
\format{
An ijtiff_img object with 890880 elements representing a 2D tomogram.
}
\source{
CZ Biohub CryoET Data Portal:
\url{https://cryoetdataportal.czscience.com/datasets/10006?ground_truth=true}
}
\usage{
TS_001.133
}
\description{
A sample tomogram in .tif format containing SARS-CoV-2 spike proteins on
intact virions.
}
\examples{
  \dontrun{
  TS_001.133
 }
}
\keyword{datasets}
