% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TS_001.133_ground_truth}
\alias{TS_001.133_ground_truth}
\title{Ground Truth Mask Of SARS-CoV-2 Spike Proteins On Intact Virions}
\format{
An ijtiff_img object with 890880 elements representing a mask on
the 2D tomogram.
}
\usage{
TS_001.133_ground_truth
}
\description{
A sample ground truth mask in .tif format for the tomogram containing
SARS-CoV-2 spike proteins on intact virions.
}
\examples{
  \dontrun{
  TS_001.133_ground_truth
 }
}
\keyword{datasets}
