% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runSegmentation.R
\name{.compareGroundTruth}
\alias{.compareGroundTruth}
\title{Compare Segmentation Results Against Ground Truth}
\usage{
.compareGroundTruth(seg_result, ground_truth)
}
\arguments{
\item{seg_result}{list containing segmentation masks from different methods}

\item{ground_truth}{ground truth tomogram image read into R}
}
\value{
Returns a ggplot2 bar plot visualising the accuracy of each
segmentation method compared to the ground truth
}
\description{
This function compares the segmentation output from multiple thresholding
methods against a provided ground truth mask (also in .tif format),
using the DICE similarity coefficient as a measure of accuracy. As
the segmentation output is required for comparison, this function is
private and only used within the segmentation pipeline.
}
\references{
Dice, L. R. (1945). Measures of the amount of ecological association between
species. Ecology, 26(3), 297-302. https://doi.org/10.2307/1932409
Wickham, H. (2016). ggplot2: Elegant graphics for data analysis.
Springer-Verlag New York. https://ggplot2.tidyverse.org
}
