% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runDenoising.R
\name{runDenoising}
\alias{runDenoising}
\title{Apply Denoising Algorithms To Tomograms}
\usage{
runDenoising(image, output_dir = NULL, gaussian_sigma = 1, median_size = 5)
}
\arguments{
\item{image}{tomogram file to be denoised}

\item{output_dir}{optional directory to save the denoised images, otherwise
images will be saved in the inst/extdata folder of the package}

\item{gaussian_sigma}{optional standard deviation for Gaussian blur, default
set to 1}

\item{median_size}{optional window size of the median filter, default set
to 5}
}
\value{
Returns null, instead saves the denoised images.
}
\description{
This function performs image denoising on tomograms (in .tif format) using
several filtering algorithms implemented in the imager package. The supported
denoising algorithms include Gaussian blur, median filter, and anisotropic
diffusion filtering. The function reads and displays the input image, applies
the denoising algorithms, and visualises the results for comparison. This is
meant to assist the user in evaluating the effectiveness of different
denoising algorithms before passing them into segmentation.
}
\examples{
# Example 1:
# Running denoising on a sample tomogram

image <- TS_001.133
cryoCompare::runDenoising(image, tempdir())

}
\references{
Pau, G., Fuchs, F., Sklyar, O., Boutros, M., & Huber W. (2010).
EBImage - an R package for image processing with applications to
cellular phenotypes. Bioinformatics, 26(7), 979-981.
https://doi.org/10.1093/bioinformatics/btq046
Huang, T., Yang, G., & Tang, G. (1979). A fast two-dimensional median
filtering algorithm. IEEE Transactions on Acoustics, Speech, and Signal
Processing, 27(1), 13-18. https://doi.org/10.1109/TASSP.1979.1163188
Marr, D., & Hildreth, E. (1980). Theory of edge detection. Proceedings of
the Royal Society B: Biological Sciences, 207(1167): 187-217.
https://doi.org/10.1098/rspb.1980.0020
Nolan, R., & Padilla-Parra, S. (2018). ijtiff: An R package providing TIFF
I/O for ImageJ users. Journal of Open Source Software, 3(23), 633.
https://doi.org/10.21105/joss.00633
}
