% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runDenoising.R
\name{runDenoising}
\alias{runDenoising}
\title{Apply Denoising Algorithms To Tomograms}
\usage{
runDenoising(image_path, output_dir = NULL)
}
\arguments{
\item{image_path}{path to the tomogram file in .tif format}
}
\value{
Returns null, instead saves the denoised images.
}
\description{
This function performs image denoising on tomograms (in .tif format) using
several filtering algorithms implemented in the imager package. The supported
denoising algorithms include Gaussian blur, median filter, and anisotropic
diffusion filtering. The function reads and displays the input image, applies
the denoising algorithms, and visualises the results for comparison. This is
meant to assist the user in evaluating the effectiveness of different
denoising algorithms before passing them into segmentation.
}
\examples{
# Example 1:
# Running denoising on a sample tomogram
image_path <- system.file("data", "TS_001.133.tif", package = "cryoCompare")
denoised_results <- cryoCompare::runDenoising(image_path, tempdir())

}
\references{
https://cran.r-project.org/web/packages/ijtiff/index.html
https://cran.r-project.org/web/packages/imager/index.html
https://pubmed.ncbi.nlm.nih.gov/6102765/
https://www.sid.ir/paper/544308/en
https://www.mia.uni-saarland.de/weickert/Papers/book.pdf
}
