% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runSegmentation.R
\name{runSegmentation}
\alias{runSegmentation}
\title{Run Segmentation Pipeline On Tomograms}
\usage{
runSegmentation(image, methods, ground_truth = NULL)
}
\arguments{
\item{image}{tomogram file to be run through the pipeline}

\item{methods}{vector of strings of segmentation methods to apply; choose
from "Huang", "Mean", "Otsu", and "Triangle"}

\item{ground_truth}{optional ground truth mask of tomogram file}
}
\value{
Returns null as this is the end of the pipeline
}
\description{
This function will run segmentation algorithms on the specified tomograms
(must be in .tif format). It is preferred that the tomograms are already
denoised, as segmentation algorithms typically perform better on denoised
data. The function supports multiple segmentation algorithms, including
various traditional image thresholding methods including Huang (), Mean (),
Otsu (), and Triangle (). This function will read the tiff tomogram file,
then send the results to other functions to apply the specified segmentation
algorithms, visualise the results, and compare them against ground truth
if provided.
}
\examples{
# Example 1:
# Segmentation of a tomogram with Otsu and Triangle algorithms without ground truth
image <- TS_001.133
cryoCompare::runSegmentation(image, methods = c("Otsu", "Triangle"))

# Example 2:
# Segmentation of a tomogram with all available methods with ground truth
image <- TS_001.133
ground_truth <- TS_001.133_ground_truth
cryoCompare::runSegmentation(image,
methods = c("Huang", "Mean", "Otsu", "Triangle"), ground_truth)

}
\references{
https://www.sciencedirect.com/science/article/abs/pii/0031320394E0043K?via%3Dihub
https://www.sciencedirect.com/science/article/abs/pii/S1049965283710400?via%3Dihub
https://ieeexplore.ieee.org/document/4310076
https://journals.sagepub.com/doi/10.1177/25.7.70454
https://cran.r-project.org/web/packages/ijtiff/index.html
}
