% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runSegmentation.R
\name{viewSegmentation}
\alias{viewSegmentation}
\title{Visualise Segmentation Results}
\usage{
viewSegmentation(image, methods)
}
\arguments{
\item{image}{tomogram image read into R}

\item{methods}{vector of strings of segmentation methods to visualise}
}
\value{
Returns null as this function is used for visualisation only.
}
\description{
This function will display the mask from a specific thresholding algorithm
applied to a given tomogram, allowing users to visually inspect what each
individual segmentation method is picking up. It is primarily used within
the segmentation pipeline to visualise the results of different segmentation
methods but can also be used independently for visual analysis.
}
\examples{
#' # Example 1:
# Visualising segmentation results using Otsu and Triangle methods
image <- TS_001.133
cryoCompare::viewSegmentation(image, methods = c("Otsu", "Triangle"))

}
\references{
Landini, G., Randell, D. A., Fouad, S., & Galton, A. (2017). Automatic
thresholding from the gradients of region boundaries. Journal of Microscopy,
265(2), 185-195. https://doi.org/10.1111/jmi.12474
Nolan, R., & Padilla-Parra, S. (2018). ijtiff: An R package providing TIFF
I/O for ImageJ users. Journal of Open Source Software, 3(23), 633.
https://doi.org/10.21105/joss.00633
}
